requirements.txt
django
bs4
requests
celery
redis
pytz
python-dotenv
